requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
urllib3==1.25.9
